move the functionalities in the baseparser.py to base/parser.py and base/serializer.py
