+++ hydrolib/core/base/parser.py
@@ -1,4 +1,5 @@
-from typing import List, Tuple
+from typing import List, Tuple, Dict
+from pathlib import Path
 
 
 class BaseParser:
@@ -39,3 +40,9 @@ class BaseParser:
             raise ValueError(
                 f"Line {line_index}: comments are only supported at the start of the file, before the data."
             )
+
+
+class DummmyParser:
+    @staticmethod
+    def parse(filepath: Path) -> Dict:
+        return {}
