+++ hydrolib/core/base/serializer.py
@@ -2,11 +2,6 @@ from pathlib import Path
 from typing import Dict
 
 
-class DummmyParser:
-    @staticmethod
-    def parse(filepath: Path) -> Dict:
-        return {}
-
 
 class DummySerializer:
     @staticmethod
